rename: 함수/변수명 (expiration > expirationDate)
