+++ src/hooks/creditCard/index.ts
@@ -1,3 +1,3 @@
-export { default as useChangeExpirationDate } from './useChangeExpiration';
+export { default as useChangeExpirationDate } from './useChangeExpirationDate';
 export { default as useChangeCardNumber } from './useChangeCardNumber';
 export { default as useChangeOwnerName } from './useChangeOwnerName';
